Some fixes with error reporting

+++ helpers.h
@@ -20,7 +20,7 @@ void free_dynStr(dynStr *str);
 
 void print_help(char *name);
 void reverse(char *str);
-bool is_digit(char c, unsigned int base);
+bool is_digit(char c, unsigned short base);
 bool is_unknown_char(scanner *scanner, char c);
 int char_to_dec(char c, short base, bool *ok);
 char int_to_char(int v);
